refactor(complexity): extract AST helpers + simplify workflow validator
Auto-sprint issue #155 — reduce cyclomatic complexity on D-grade files:

1. validate_workflows.py (C 43.3 → target B):
   - Extract _validate_workflow_jobs() helper (eliminates nested loops)
   - Extract _validate_trigger_inputs() helper (flattens conditional chain)

2. tree_sitter_analyzer/ast_cache.py (2392 → 1886 lines):
   - Move _has_fts5, _EXCLUDE_DIRS, _worker_index_file, _content_hash,
     all node-type frozensets, and all symbol/import/call-edge extraction
     functions to new _ast_extraction.py module
   - Re-export all symbols from ast_cache.py for backward compatibility
   - _worker_index_file stays module-level in _ast_extraction.py so
     multiprocessing.Pool can pickle it by import path
